this is my last commit, i'm exhausted. i implemented to not show replies/likes/repost of deleted tweets.
there are still a few api's i didn't get to such as the user and tweet mentions, tweet/tag, tweet/context, tag/label, and fixing the users/username/feed.
I originally had the user/username/feed working but i dont know why it's not anymore and i'm tired to debug it. there are still a few bugs in terms of avoiding duplicates when following a user or liking a tweet.